refactor : day component 분리
